pyproject.toml
[project]
name = "master"
version = "0.1.0"
description = ""
authors = [
    {name = "SindreEieLedsaak",email = "sindre.eie@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit (>=1.43.2,<2.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymongo (>=4.11.3,<5.0.0)",
    "python-gitlab (>=5.6.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
